requirements.txt
streamlit
python-dotenv
openai
langchain
pydantic
langchain-community